Validate Add / Edit Timer Form
Until now a user was able to create timers without a name and with
invalid intervals. To avoid this, we now validate the form to add / edit
a timer. This means that the timer name and interval name is always
required and that an interval must be larger then 0 seconds.