Add Option to Sort Timers (#2)
Until now the timers where shown in the order in which they were
created. Now it is also possible to sort the timers, so that they are
shown in the users prefered order.

In this commit we also fixed a bug, wwere an error was thrown when a
timer didn't had an interval. This bug occured in the reduce function to
determine the complete timer time and on the details page. When a timer
doesn't contain an interval it is not possible to view the details page
anymore.